+++ samples/compose-navigation-app/src/main/kotlin/dev/zacsweers/metro/sample/androidviewmodel/MetroApp.kt
@@ -1,0 +1,13 @@
+// Copyright (C) 2025 Zac Sweers
+// SPDX-License-Identifier: Apache-2.0
+package dev.zacsweers.metro.sample.androidviewmodel
+
+import android.app.Application
+import dev.zacsweers.metro.createGraph
+import dev.zacsweers.metro.sample.androidviewmodel.components.AppGraph
+import dev.zacsweers.metro.sample.androidviewmodel.components.MetroAppComponentFactory
+
+class MetroApp : Application() {
+  /** Holder reference for the app graph for [MetroAppComponentFactory]. */
+  val appGraph by lazy { createGraph<AppGraph>() }
+}